Each user will see their specific contacts.

+++ src/app/contact-list/contact-list.component.ts
@@ -1,34 +1,37 @@
 import { Component, OnInit } from '@angular/core';
 import {CrudService} from '../services/crud.service';
+import { AuthService } from '../services/auth.service';
 
 @Component({
   selector: 'app-contact-list',
   templateUrl: './contact-list.component.html',
   styleUrls: ['./contact-list.component.css']
 })
-export class ContactListComponent {//implements OnInit {
+export class ContactListComponent implements OnInit {
   contact: any;
   contactName: string;
   contactEmail: string;
   contactPhone: string;
   message:string;
   
-
-  constructor(public crudservice:CrudService) { }
+  constructor(private authservice: AuthService,public crudservice:CrudService) { }
   
   ngOnInit() {
-    this.crudservice.get_AllContacts().subscribe(data =>{
-      this.contact = data.map(c => {
-        return {
-          id: c.payload.doc.id,
-          isEdit: false,
-          name: c.payload.doc.data()['name'],
-          email: c.payload.doc.data()['email'],
-          phone: c.payload.doc.data()['phone'],
-        };
-      })
-      console.log(this.contact);
-    });
+    if(this.authservice.currentUser != null)//We will make sure the user is logged in
+    {
+      this.crudservice.get_AllContacts().subscribe(data => {
+        this.contact = data.map(c => {
+          return {
+            id: c.payload.doc.id,
+            isEdit: false,
+            name: c.payload.doc.data()['name'],
+            email: c.payload.doc.data()['email'],
+            phone: c.payload.doc.data()['phone'],
+          };
+        })
+        console.log(this.contact);
+      });  
+    }
   }
   
   /*CreateRecord() will fire after the user press the "Create Contact" btn*/
@@ -51,7 +54,7 @@ export class ContactListComponent {//implements OnInit {
       console.log(error);
     })
   }
-
+//-----------------------------------------------------------------
   //will fire after the user press "Edit Contant"
   editRecord(Record)
   {
@@ -70,9 +73,12 @@ export class ContactListComponent {//implements OnInit {
     this.crudservice.update_contact(recordData.id, record);//we defined update_contact() in crud.service.ts
     recordData.isEdit = false;
   }
-  
+
   //will fire after the user press "Delete Contact"
   DeleteContact(recordId){
+    if(this.authservice.currentUser != null)//We will make sure the user is logged in
+    {
     this.crudservice.delete_contact(recordId);
+    }
   }
 }
